python-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
opencv-python==4.12.0.88
openai-whisper==20250625
numpy==2.2.6
python-multipart==0.0.20
